adding useCheckIfIsLikedAndAddToFavorites hook
